- Removed parameter sets in favor of manual validation of switches
- Updated position of the Message parameter in Write-PLPLog so that an invocation of the function can be formed as (for example): "$log | Write-PLPLog 'Log this message'"

+++ cmdlets/New-PLPLog.ps1
@@ -8,11 +8,9 @@ function New-PLPLog {
         [string]
         $FilePath,
 
-        [Parameter(ParameterSetName='ParamSetOverwrite')]
         [switch]
         $Overwrite,
 
-        [Parameter(ParameterSetName='ParamSetAppend')]
         [switch]
         $Append
     )
@@ -20,6 +18,12 @@ function New-PLPLog {
     Begin {}
 
     Process {
+        # Validate switch combinations
+        if ($Overwrite -and $Append) {
+            Write-Error 'The -Overwrite and -Append parameters cannot be used together; they are mutually exclusive.'
+            return $null
+        }
+
         if (Test-Path $FilePath -PathType Leaf) {
             if ($Overwrite) {
                 Remove-Item -Path $FilePath -Force
